+++ vite-project/package.json
@@ -16,7 +16,8 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-hook-form": "^7.43.9",
-    "react-router-dom": "^6.9.0"
+    "react-router-dom": "^6.9.0",
+    "unsplash-js": "^7.0.15"
   },
   "devDependencies": {
     "@testing-library/jest-dom": "^5.16.5",
