+++ vite-project/src/types/interfaces.ts
@@ -16,3 +16,20 @@ export interface IFormCardPageState {
   formCards: IFormCard[];
   isCardHidden: boolean;
 }
+
+export interface ICarData {
+  id?: number;
+  name: string;
+  year: number;
+  color: string;
+  km: number;
+  img: string;
+  price: number;
+  description: string;
+}
+
+export interface IResult {
+  total: number;
+  total_pages: number;
+  results: ICarData;
+}
